requirements.txt
numpy
mysql-connector-python
